+++ package.json
@@ -54,6 +54,7 @@
   "dependencies": {
     "@project-serum/anchor": "^0.20.1",
     "@project-serum/serum": "^0.13.61",
+    "@saberhq/solana-contrib": "^1.12.41",
     "@saberhq/token-utils": "^1.12.37",
     "@solana/spl-token": "^0.1.8",
     "@solana/spl-token-registry": "^0.2.2086",
